@@ -3,11 +3,8 @@ name = "pytest-broadcaster"
 description = "Pytest plugin to broadcast pytest output to various destinations"
 authors = [{ name = "Guillaume Charbonnier", email = "gu.charbon@gmail.com" }]
 urls = { "Source" = "https://github.com/charbonnierg/pytest-broadcaster", "Tracker" = "https://github.com/charbonnierg/pytest-broadcaster/issues" }
-
-dependencies = ["pytest", 'tomli>=1; python_version < "3.11"']
-dynamic = ["version"]
 readme = "README.md"
-requires-python = ">= 3.8"
+requires-python = ">= 3.9"
 classifiers = [
     "Development Status :: 3 - Alpha",
     "Framework :: Pytest",
