@@ -4,16 +4,15 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
-from typing import List, Union
+from typing import TYPE_CHECKING
 
-from . import test_case, test_directory, test_module, test_suite
+if TYPE_CHECKING:
+    from . import test_case, test_directory, test_module, test_suite
 
 
 @dataclass
 class CollectReport:
-    """
-    A report of collected items.
-    """
+    """A report of collected items."""
 
     session_id: str
     """
@@ -27,13 +26,11 @@ class CollectReport:
     """
     The date and time when the report was generated in ISO 8601 format.
     """
-    items: List[
-        Union[
-            test_directory.TestDirectory,
-            test_module.TestModule,
-            test_suite.TestSuite,
-            test_case.TestCase,
-        ]
+    items: list[
+        test_directory.TestDirectory
+        | test_module.TestModule
+        | test_suite.TestSuite
+        | test_case.TestCase
     ]
     """
     An array of collected items. Each collected item is a test directory, test module, test suite, or test case. Top level directory is always first element in the array, followed by test cases, then test suites, then test modules, then test directories for each level of nesting.
