@@ -4,16 +4,15 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
-from typing import Optional
+from typing import TYPE_CHECKING
 
-from . import outcome, test_case_error
+if TYPE_CHECKING:
+    from . import outcome, test_case_error
 
 
 @dataclass
 class TestCaseSetup:
-    """
-    Pytest Test Case Setup
-    """
+    """Pytest Test Case Setup."""
 
     session_id: str
     """
@@ -43,7 +42,7 @@ class TestCaseSetup:
     """
     The event type. Always set to 'case_setup'.
     """
-    error: Optional[test_case_error.TestCaseError] = None
+    error: test_case_error.TestCaseError | None = None
     """
     Error details if the setup step failed.
     """
