@@ -18,6 +18,7 @@ SOURCES = Path(__file__).parent.parent.joinpath("src/schemas")
 
 
 def replace_refs_and_ids(schema: dict[str, Any], ref_prefix: str) -> None:
+    """Replace host in $ref and $id fields in given JSON schema definition."""
     if "$ref" in schema:
         ref_template: str = schema["$ref"]
         ref_value = ref_template.split("/")[0]
@@ -35,18 +36,22 @@ def replace_refs_and_ids(schema: dict[str, Any], ref_prefix: str) -> None:
 
 
 def create_schema(source: Path, output_directory: Path, ref_prefix: str) -> None:
+    """Create JSON schema in output directory given some source file and prefix."""
     destination = output_directory.joinpath(source.name)
     content = json.loads(source.read_text())
     replace_refs_and_ids(content, ref_prefix)
     destination.write_text(json.dumps(content, indent=2))
 
 
-def generate_schemas(version: str, repository: str, output_directory: str):
+def generate_schemas(version: str, repository: str, output_directory: str) -> None:
+    """Generate JSON schemas for a given version and git repository."""
     # Validate reference and output
     if not (version and repository and output_directory):
-        raise ValueError("version, repository and output_directory are required")
+        error = "version, repository and output_directory must not be empty"
+        raise ValueError(error)
     if output_directory.startswith("/"):
-        raise ValueError("output must be relative")
+        error = "output must be a relative relative path"
+        raise ValueError(error)
     ref_prefix = repository
     if not ref_prefix.endswith("/"):
         ref_prefix += "/"
@@ -59,7 +64,7 @@ def generate_schemas(version: str, repository: str, output_directory: str):
     # Normalize repository and output
     if not ref_prefix.endswith("/"):
         ref_prefix += "/"
-    if output_directory.endswith("/"):
+    if output_directory.endswith("/"):  # noqa: FURB188
         output_directory = output_directory[:-1]
     # Build full repository URL
     target = output_directory
@@ -104,20 +109,20 @@ def generate_schemas_index(schemas: dict[str, dict[str, Any]]) -> None:
         content += "| ------ | ----------- | --- |"
         content += "\n"
         for schema_name, schema in schemas.items():
-            schema_description = schema["description"]
+            schema_desc = schema["description"]
             schema_url = schema["url"]
-            content += f"| {schema_name} | {schema_description} | [{schema_url}]({schema_url}) |"
+            row = f"| {schema_name} | {schema_desc} | [{schema_url}]({schema_url}) |"
+            content += row
             content += "\n"
         index.write(content)
 
 
 def generate_license_file() -> None:
     """Generate the LICENSE.md file."""
-
-    with mkdocs_gen_files.open("LICENSE.md", "w") as license_file:
-        license = Path("LICENSE")
-        if license.is_file():
-            content = license.read_text()
+    with mkdocs_gen_files.open("LICENSE.md", "w") as docs_license_file:
+        repo_license_file = Path("LICENSE")
+        if repo_license_file.is_file():
+            content = repo_license_file.read_text()
             content += "\n\n"
             content += """
 <style>
@@ -126,12 +131,11 @@ def generate_license_file() -> None:
   }
 </style>
 """
-            license_file.write(content)
+            docs_license_file.write(content)
         else:
-            license_file.write("No license file found")
+            docs_license_file.write("No license file found")
 
 
-# generate_references()
 generate_license_file()
 generate_schemas(
     version=os.environ.get("VERSION", "latest"),
