@@ -1,3 +1,5 @@
+"""pytest_broadcaster package."""
+
 from .__about__ import __version__, __version_tuple__
 from ._internal._json_files import JSONFile, JSONLinesFile
 from ._internal._reporter import DefaultReporter
@@ -5,12 +7,12 @@ from ._internal._webhook import HTTPWebhook
 from .interfaces import Destination, Reporter
 
 __all__ = [
-    "__version__",
-    "__version_tuple__",
-    "Destination",
-    "Reporter",
     "DefaultReporter",
+    "Destination",
     "HTTPWebhook",
     "JSONFile",
     "JSONLinesFile",
+    "Reporter",
+    "__version__",
+    "__version_tuple__",
 ]
