@@ -1,5 +1,5 @@
 # generated by datamodel-codegen:
-#   filename:  session_finish.json
+#   filename:  session_end.json
 
 from __future__ import annotations
 
@@ -7,10 +7,8 @@ from dataclasses import dataclass
 
 
 @dataclass
-class SessionFinish:
-    """
-    Event emitted when the session is finished.
-    """
+class SessionEnd:
+    """Event emitted when the session is finished."""
 
     session_id: str
     """
@@ -24,7 +22,7 @@ class SessionFinish:
     """
     The status which pytest will return to the system. 0 means success, 1 means test failure, anything else is an error.
     """
-    event: str = "session_finish"
+    event: str = "session_end"
     """
-    The event type. Always set to 'session_finish'.
+    The event type. Always set to 'session_end'.
     """
