@@ -1,18 +1,18 @@
 # generated by datamodel-codegen:
-#   filename:  test_case_finished.json
+#   filename:  test_case_end.json
 
 from __future__ import annotations
 
 from dataclasses import dataclass
+from typing import TYPE_CHECKING
 
-from . import outcome
+if TYPE_CHECKING:
+    from . import outcome
 
 
 @dataclass
-class TestCaseFinished:
-    """
-    Event emitted when a test case is finished (meaning all of setup, call and teardown steps are finished).
-    """
+class TestCaseEnd:
+    """Event emitted when a test case is finished (meaning all of setup, call and teardown steps are finished)."""
 
     session_id: str
     """
@@ -38,7 +38,7 @@ class TestCaseFinished:
     """
     Outcome of the test case.
     """
-    event: str = "case_finished"
+    event: str = "case_end"
     """
-    The event type. Always set to 'case_finished'.
+    The event type. Always set to 'case_end'.
     """
