@@ -1,12 +1,15 @@
 from __future__ import annotations
 
 import sys
-from pathlib import Path
+from typing import TYPE_CHECKING
 
 import pytest
 
 from _testing.setup import CommonTestSetup
 from pytest_broadcaster import __version__
+
+if TYPE_CHECKING:
+    from pathlib import Path
 
 
 @pytest.mark.basic
@@ -29,7 +32,6 @@ class TestBasicSkip(CommonTestSetup):
 
     def test_json(self):
         """Test JSON report for single test case within single test file."""
-
         directory = self.make_test_directory()
         result = self.test_dir.runpytest("--collect-report", self.json_file.as_posix())
         assert result.ret == 0
@@ -82,7 +84,7 @@ class TestBasicSkip(CommonTestSetup):
                         "error": None,
                     },
                     "finished": {
-                        "event": "case_finished",
+                        "event": "case_end",
                         "session_id": "omitted",
                         "node_id": "test_basic_skipped.py::test_skipped",
                         "outcome": "skipped",
@@ -152,7 +154,6 @@ class TestBasicSkip(CommonTestSetup):
 
     def test_jsonl_basic(self):
         """Test JSON Lines report for single test case within single test file."""
-
         directory = self.make_test_directory()
         result = self.test_dir.runpytest(
             "--collect-log", self.json_lines_file.as_posix()
@@ -254,7 +255,7 @@ class TestBasicSkip(CommonTestSetup):
                 "error": None,
             },
             {
-                "event": "case_finished",
+                "event": "case_end",
                 "session_id": "omitted",
                 "node_id": "test_basic_skipped.py::test_skipped",
                 "outcome": "skipped",
@@ -264,7 +265,7 @@ class TestBasicSkip(CommonTestSetup):
             },
             {
                 "exit_status": 0,
-                "event": "session_finish",
+                "event": "session_end",
                 "session_id": "omitted",
                 "timestamp": "omitted",
             },
