@@ -1,17 +1,20 @@
 from __future__ import annotations
 
 import sys
-from pathlib import Path
+from typing import TYPE_CHECKING
 
 import pytest
 
 from _testing.setup import CommonTestSetup
 from pytest_broadcaster import __version__
 
+if TYPE_CHECKING:
+    from pathlib import Path
+
 
 @pytest.mark.basic
 class TestBasicxFail(CommonTestSetup):
-    """Scenario: A single test case within a single test file which is marked with xfail."""
+    """Test single test case within a single test file which is marked with xfail."""
 
     def make_test_directory(self) -> Path:
         self.make_testfile(
@@ -39,7 +42,6 @@ class TestBasicxFail(CommonTestSetup):
 
     def test_json(self):
         """Test JSON report for single test case within single test file."""
-
         directory = self.make_test_directory()
         result = self.test_dir.runpytest("--collect-report", self.json_file.as_posix())
         assert result.ret == 0
@@ -101,7 +103,7 @@ class TestBasicxFail(CommonTestSetup):
                         "error": None,
                     },
                     "finished": {
-                        "event": "case_finished",
+                        "event": "case_end",
                         "session_id": "omitted",
                         "node_id": "test_basic.py::test_ok",
                         "outcome": "passed",
@@ -145,7 +147,7 @@ class TestBasicxFail(CommonTestSetup):
                         "error": None,
                     },
                     "finished": {
-                        "event": "case_finished",
+                        "event": "case_end",
                         "session_id": "omitted",
                         "node_id": "test_basic_xfail.py::test_xfail",
                         "outcome": "xfailed",
@@ -247,7 +249,6 @@ class TestBasicxFail(CommonTestSetup):
 
     def test_jsonl(self):
         """Test JSON Lines report for single test case within single test file."""
-
         directory = self.make_test_directory()
         result = self.test_dir.runpytest(
             "--collect-log", self.json_lines_file.as_posix()
@@ -391,7 +392,7 @@ class TestBasicxFail(CommonTestSetup):
                 "error": None,
             },
             {
-                "event": "case_finished",
+                "event": "case_end",
                 "session_id": "omitted",
                 "node_id": "test_basic.py::test_ok",
                 "outcome": "passed",
@@ -430,7 +431,7 @@ class TestBasicxFail(CommonTestSetup):
                 "error": None,
             },
             {
-                "event": "case_finished",
+                "event": "case_end",
                 "session_id": "omitted",
                 "node_id": "test_basic_xfail.py::test_xfail",
                 "outcome": "xfailed",
@@ -440,7 +441,7 @@ class TestBasicxFail(CommonTestSetup):
             },
             {
                 "exit_status": 0,
-                "event": "session_finish",
+                "event": "session_end",
                 "session_id": "omitted",
                 "timestamp": "omitted",
             },
